requirements.txt
Django==4.2
djangorestframework==3.14
requests==2.31
djangorestframework-simplejwt==5.0
gunicorn==21.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
h11==0.14.0
idna==3.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
